statement: rename prepared_statement.rs -> prepared.rs
