+++ docs/source/retry-policy/fallthrough.md
@@ -63,7 +63,7 @@ To use in a [prepared query](../queries/prepared.md):
 # use std::error::Error;
 # use std::sync::Arc;
 # async fn check_only_compiles(session: &Session) -> Result<(), Box<dyn Error>> {
-use scylla::statement::prepared_statement::PreparedStatement;
+use scylla::statement::prepared::PreparedStatement;
 use scylla::client::execution_profile::ExecutionProfile;
 use scylla::policies::retry::FallthroughRetryPolicy;
