+++ docs/source/tracing/paged.md
@@ -50,7 +50,7 @@ for id in tracing_ids {
 # use scylla::client::session::Session;
 # use std::error::Error;
 # async fn check_only_compiles(session: &Session) -> Result<(), Box<dyn Error>> {
-use scylla::statement::prepared_statement::PreparedStatement;
+use scylla::statement::prepared::PreparedStatement;
 use scylla::observability::tracing::TracingInfo;
 use futures::StreamExt;
 use uuid::Uuid;
