+++ scylla/src/client/caching_session.rs
@@ -3,7 +3,7 @@ use crate::response::query_result::QueryResult;
 use crate::response::{PagingState, PagingStateResponse};
 use crate::routing::partitioner::PartitionerName;
 use crate::statement::batch::{Batch, BatchStatement};
-use crate::statement::prepared_statement::PreparedStatement;
+use crate::statement::prepared::PreparedStatement;
 use crate::statement::query::Query;
 use bytes::Bytes;
 use dashmap::DashMap;
@@ -253,7 +253,7 @@ mod tests {
     use crate::response::PagingState;
     use crate::routing::partitioner::PartitionerName;
     use crate::statement::batch::{Batch, BatchStatement};
-    use crate::statement::prepared_statement::PreparedStatement;
+    use crate::statement::prepared::PreparedStatement;
     use crate::statement::query::Query;
     use crate::test_utils::{
         create_new_session_builder, scylla_supports_tablets, setup_tracing, PerformDDL,
