+++ scylla/src/cluster/state.rs
@@ -5,7 +5,7 @@ use crate::routing::locator::tablets::{RawTablet, Tablet, TabletsInfo};
 use crate::routing::locator::ReplicaLocator;
 use crate::routing::partitioner::{calculate_token_for_partition_key, PartitionerName};
 use crate::routing::{Shard, Token};
-use crate::statement::prepared_statement::TokenCalculationError;
+use crate::statement::prepared::TokenCalculationError;
 
 use itertools::Itertools;
 use scylla_cql::frame::response::result::TableSpec;
