+++ scylla/tests/integration/silent_prepare_batch.rs
@@ -1,7 +1,7 @@
 use crate::utils::{create_new_session_builder, setup_tracing, unique_keyspace_name, PerformDDL};
 use scylla::client::session::Session;
 use scylla::statement::batch::Batch;
-use scylla::statement::prepared_statement::PreparedStatement;
+use scylla::statement::prepared::PreparedStatement;
 
 use std::collections::BTreeSet;
