+++ scylla/tests/integration/tablets.rs
@@ -18,7 +18,7 @@ use scylla::policies::load_balancing::LoadBalancingPolicy;
 use scylla::policies::load_balancing::RoutingInfo;
 use scylla::response::query_result::QueryResult;
 use scylla::serialize::row::SerializeRow;
-use scylla::statement::prepared_statement::PreparedStatement;
+use scylla::statement::prepared::PreparedStatement;
 use scylla::statement::query::Query;
 
 use scylla::errors::ExecutionError;
